// src/constants/construct-tree-context-types.ts
export enum ConstructTreeContextTypes {
  TagContent = "TagContent",
  Tag = "Tag",
  TagName = "TagName",
  Attributes = "Attributes",
  Attribute = "Attribute",
  AttributeValue = "AttributeValue",
  Comment = "Comment",
  Doctype = "Doctype",
  DoctypeAttributes = "DoctypeAttributes",
  DoctypeAttribute = "DoctypeAttribute",
  ScriptTag = "ScriptTag",
  StyleTag = "StyleTag",
}